feat: 스프린트 수정

+++ src/main/java/com/official/memento/schedule/infrastructure/persistence/projection/ScheduleOrderInfoProjection.java
@@ -15,7 +15,7 @@ public interface ScheduleOrderInfoProjection {
     LocalDateTime getEndDate();
     boolean getIsAllDay();
     ScheduleType getType();
-    Integer getOrderNum();
+    Double getOrderNum();
     String getTagName();
     TagColor getTagColor();
     Long getTagId();
